backshots/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-dotenv
pydantic
